+++ include/compute/arith/ArithMuxExecutor.h
@@ -11,7 +11,7 @@ private:
     bool _cond_i{};
 
 public:
-    ArithMuxExecutor(int64_t x, int64_t y, bool c, int l, int32_t objTag, int8_t msgTagOffset, int clientRank);
+    ArithMuxExecutor(int64_t x, int64_t y, bool c, int l, int16_t objTag, int16_t msgTagOffset, int clientRank);
 
     ArithMuxExecutor *execute() override;
